+++ src/microservices/Microservices.IsIdentifiable/Service/Classifier.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.IO.Abstractions;
 using System.Linq;
 using Microservices.IsIdentifiable.Reporting.Reports;
 
@@ -19,7 +20,7 @@ namespace Microservices.IsIdentifiable.Service
                 throw new DirectoryNotFoundException($"Could not find directory {DataDirectory.FullName}");
         }
 
-        public abstract IEnumerable<Reporting.Failure> Classify(FileInfo dcm);
+        public abstract IEnumerable<Reporting.Failure> Classify(IFileInfo dcm);
 
         /// <summary>
         /// Finds a single directory of a given name in the <see cref="DataDirectory"/> and asserts that it exists