+++ src/microservices/Microservices.IsIdentifiable/Service/TesseractStanfordDicomFileClassifier.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.IO.Abstractions;
 using System.Linq;
 using Microservices.IsIdentifiable.Options;
 using Microservices.IsIdentifiable.Reporting.Reports;
@@ -13,13 +14,15 @@ namespace Microservices.IsIdentifiable.Service
         private DicomFileRunner _runner;
 
         //public TesseractStanfordDicomFileClassifier(DirectoryInfo dataDirectory) : base(dataDirectory)
-        public TesseractStanfordDicomFileClassifier(DirectoryInfo dataDirectory) : base(dataDirectory)
+        public TesseractStanfordDicomFileClassifier(DirectoryInfo dataDirectory, IsIdentifiableServiceOptions isIdentifiableServiceOptions) : base(dataDirectory)
         {
             var fileOptions = new IsIdentifiableDicomFileOptions();
             
             //need to pass this so that the runner doesn't get unhappy about there being no reports (even though we clear it below)
             fileOptions.ColumnReport = true;
             fileOptions.TessDirectory = dataDirectory.FullName;
+
+            fileOptions.IgnoreTextLessThan = isIdentifiableServiceOptions.IgnoreTextLessThan;
 
             // The Rules directory is always called "IsIdentifiableRules"
             DirectoryInfo[] subDirs = dataDirectory.GetDirectories("IsIdentifiableRules");
@@ -31,7 +34,7 @@ namespace Microservices.IsIdentifiable.Service
 
         
 
-        public override IEnumerable<Reporting.Failure> Classify(FileInfo dcm)
+        public override IEnumerable<Reporting.Failure> Classify(IFileInfo dcm)
         {
             _runner.Reports.Clear();
             var toMemory = new ToMemoryFailureReport();