+++ tests/microservices/Microservices.IsIdentifiable.Tests/ServiceTests/TesseractStanfordDicomFileClassifierTests.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
+using Microservices.IsIdentifiable.Options;
 using Microservices.IsIdentifiable.Service;
 using NUnit.Framework;
 
@@ -15,7 +16,7 @@ namespace Microservices.IsIdentifiable.Tests.ServiceTests
         public void TestDataDirectory_DoesNotExist()
         {
             var d = new DirectoryInfo("asdflsdfjadfshsdfdsafldsf;dsfldsafj");
-            Assert.Throws<DirectoryNotFoundException>(()=>new TesseractStanfordDicomFileClassifier(d));
+            Assert.Throws<DirectoryNotFoundException>(()=>new TesseractStanfordDicomFileClassifier(d, new IsIdentifiableServiceOptions()));
         }
         [Test]
         public void TestDataDirectory_Empty()
@@ -24,7 +25,7 @@ namespace Microservices.IsIdentifiable.Tests.ServiceTests
 
             var d = new DirectoryInfo(path);
             d.Create();
-            Assert.Throws<FileNotFoundException>(()=>new TesseractStanfordDicomFileClassifier(d));
+            Assert.Throws<FileNotFoundException>(()=>new TesseractStanfordDicomFileClassifier(d, new IsIdentifiableServiceOptions()));
         }
     }
 }